Migrate wishlist REST API to GraphQL; remove deprecated REST endpoints
Replaced the wishlist REST API with a new GraphQL implementation for better flexibility and schema-driven development. Removed `StorefrontWishlistResource` and associated hydration services. Introduced `WishlistResource` GraphQL API, integrating with existing services, and added comprehensive unit tests for query/mutation coverage. Updated `PriceUtils` return types for consistency.

+++ src/main/java/org/ecommerce/backend/mapper/ProductMapper.java
@@ -113,7 +113,7 @@ public interface ProductMapper
         return productDto;
     }
 
-    default Long calculateSaleDaysRemaining(VariantPricesEntity variantPricesEntity)
+    default Integer calculateSaleDaysRemaining(VariantPricesEntity variantPricesEntity)
     {
         if (variantPricesEntity == null) {
             return null;
